Extract KORE implementation conversion
